Move PaywallComponentViewModel to own file
